CIRC-1091 remove renewal strategies
